refactor(ui): split status-line assembly into per-control helpers
Stream D of the readability branch — zero behavior/UI change. The 458-line
status_line_parts_with_labels_reusing moves to control_box/status_line.rs,
decomposed into 8 per-control push_* helpers (playback state, queue position,
rating, live transport, shuffle/repeat, EQ, streaming mode, download tag).

Bodies moved verbatim: adversarial review reconstructed the helpers inline and
the normalized result is md5-identical to the original — parts.push order,
guards, Cow variants, and every ko/ja string byte-equal, so hit-rect
alignment (EAW-sensitive) is untouched. Helper seams are &mut Vec/&App/Copy
params only; no new allocations in the render-frame path.

Sizes: control_box.rs 1536->1079, status_line.rs 560.
